refactor(server): test 共通ヘルパを __tests__/helpers/ に移動し biome の test 向け設定を整理
- server/src/test-helpers/ を server/src/__tests__/helpers/ へ移動し、各 test の
  import を `~/__tests__/helpers/...` の alias 形式に統一する
- biome.json の overrides で test ファイル向けに useArrowFunction を off にし、
  vi.fn(class) constructor mock 用の個別 biome-ignore を撤去
- vitest.config.ts の coverage exclude と CLAUDE.md の共通ヘルパ説明を更新
