+++ server/src/mastra/tools/admin-emergency-tool.test.ts
@@ -11,7 +11,7 @@ const { emergencyRepository } = await import(
 );
 const { adminEmergencyTool } = await import("./admin-emergency-tool");
 
-import { callTool } from "../../test-helpers/tool-context";
+import { callTool } from "~/__tests__/helpers/tool-context";
 
 const fakeDb = {} as D1Database;
 const adminUser = { id: "u-1", role: "admin" as const };
