+++ server/src/mastra/tools/broadcast-get-tool.test.ts
@@ -13,7 +13,7 @@ const { broadcastRepository } = await import(
 );
 const { broadcastGetTool } = await import("./broadcast-get-tool");
 
-import { callTool } from "../../test-helpers/tool-context";
+import { callTool } from "~/__tests__/helpers/tool-context";
 
 const fakeDb = {} as D1Database;
 const dbValues = { db: fakeDb };
