+++ server/src/mastra/tools/emergency-update-tool.test.ts
@@ -12,7 +12,7 @@ const { emergencyRepository } = await import(
 );
 const { emergencyUpdateTool } = await import("./emergency-update-tool");
 
-import { callTool } from "../../test-helpers/tool-context";
+import { callTool } from "~/__tests__/helpers/tool-context";
 
 const fakeDb = {} as D1Database;
 const dbValues = { db: fakeDb };
