+++ server/src/mastra/tools/persona-get-tool.test.ts
@@ -9,7 +9,7 @@ vi.mock("~/repository/persona-repository", () => ({
 const { personaRepository } = await import("~/repository/persona-repository");
 const { personaGetTool } = await import("./persona-get-tool");
 
-import { callTool } from "../../test-helpers/tool-context";
+import { callTool } from "~/__tests__/helpers/tool-context";
 
 const fakeDb = {} as D1Database;
 const adminUser = { id: "u-1", role: "admin" as const };
