+++ server/src/mastra/tools/poll-get-tool.test.ts
@@ -15,7 +15,7 @@ const { pollRepository } = await import("~/repository/poll-repository");
 const { getPollResults } = await import("~/services/poll-response");
 const { pollGetTool } = await import("./poll-get-tool");
 
-import { callTool } from "../../test-helpers/tool-context";
+import { callTool } from "~/__tests__/helpers/tool-context";
 
 const fakeDb = {} as D1Database;
 const dbValues = { db: fakeDb };
