+++ server/src/mastra/tools/village-search-tool.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from "vitest";
-import { callTool } from "../../test-helpers/tool-context";
+import { callTool } from "~/__tests__/helpers/tool-context";
 import { villageSearchTool } from "./village-search-tool";
 
 describe("villageSearchTool.execute", () => {
