+++ server/src/middleware/require-thread-access.test.ts
@@ -11,7 +11,6 @@ vi.mock("~/lib/storage", () => ({
 }));
 
 vi.mock("@mastra/memory", () => ({
-  // biome-ignore lint/complexity/useArrowFunction: new で呼ばれる constructor mock
   Memory: vi.fn(function () {
     return {
       getThreadById: mockGetThreadById,
