+++ server/src/repository/user-poll-state-repository.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-import { createTestDb, type TestDb } from "../test-helpers/test-db";
+import { createTestDb, type TestDb } from "~/__tests__/helpers/test-db";
 
 const { testDbHolder } = vi.hoisted(() => ({
   testDbHolder: { db: null as TestDb | null },
