+++ server/src/routes/admin/invitations.test.ts
@@ -36,7 +36,7 @@ const { adminUserRepository } = await import(
 );
 const { invitationRoutes: rawRoutes } = await import("./invitations");
 
-import { withResolvePrincipal } from "../../test-helpers/test-app";
+import { withResolvePrincipal } from "~/__tests__/helpers/test-app";
 
 const routes = await withResolvePrincipal(rawRoutes);
