+++ server/src/routes/admin/poll.test.ts
@@ -48,7 +48,7 @@ const { adminUserRepository } = await import(
 );
 const { pollAdminRoutes: rawRoutes } = await import("./poll");
 
-import { withResolvePrincipal } from "../../test-helpers/test-app";
+import { withResolvePrincipal } from "~/__tests__/helpers/test-app";
 
 const routes = await withResolvePrincipal(rawRoutes);
