+++ server/src/routes/auth.anonymous-session.test.ts
@@ -38,7 +38,7 @@ vi.mock("~/repository/admin-session-repository", () => ({
 
 const { authRoutes: rawAuthRoutes } = await import("./auth");
 
-import { withResolvePrincipal } from "../test-helpers/test-app";
+import { withResolvePrincipal } from "~/__tests__/helpers/test-app";
 
 const authRoutes = await withResolvePrincipal(rawAuthRoutes);
