+++ server/src/routes/threads/chat.test.ts
@@ -39,7 +39,6 @@ vi.mock("~/mastra/request-context", () => ({
 }));
 
 vi.mock("@mastra/memory", () => ({
-  // biome-ignore lint/complexity/useArrowFunction: new で呼ばれる constructor mock
   Memory: vi.fn(function () {
     return {
       listThreads: vi.fn(),
@@ -73,7 +72,7 @@ vi.mock("~/services/auth/anonymous-session", () => ({
 const sessionService = await import("~/services/auth/anonymous-session");
 const { threadsRoutes: rawThreadsRoutes } = await import("./index");
 
-import { withResolvePrincipal } from "../../test-helpers/test-app";
+import { withResolvePrincipal } from "~/__tests__/helpers/test-app";
 
 const routes = await withResolvePrincipal(rawThreadsRoutes);
