+++ server/src/services/user-deletion.test.ts
@@ -14,7 +14,7 @@ import {
   userPollState,
 } from "~/db";
 import { hmacSha256 } from "~/lib/crypto";
-import { createTestDb, type TestDb } from "../test-helpers/test-db";
+import { createTestDb, type TestDb } from "~/__tests__/helpers/test-db";
 
 const { testDbHolder } = vi.hoisted(() => ({
   testDbHolder: { db: null as TestDb | null },
